requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
starlette==0.41.3
psycopg2-binary==2.9.10
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.1.0
openpyxl==3.1.2
python-dateutil==2.8.2
reportlab==4.0.7
pytz==2023.3
